Split the editor into separate files
Scrolling through the editor code was becoming unwieldy.
