+++ thrimbletrimmer/src/editor/CategoryNotes.module.scss
@@ -1,0 +1,6 @@
+.categoryNotes {
+	border: 1px solid #666;
+	background-color: #500;
+	padding: 2px;
+	white-space: pre;
+}
