+++ thrimbletrimmer/src/editor/ChapterToggle.module.scss
@@ -1,0 +1,3 @@
+.chaptersEnabledSelection {
+	margin-top: 5px;
+}
