+++ thrimbletrimmer/src/editor/Editor.module.scss
@@ -21,85 +21,3 @@
 .errorRemoveLink {
 	float: right;
 }
-
-.clipBar {
-	width: calc(100% - 14px);
-	min-height: 7px;
-	background-color: #bbb;
-	position: relative;
-	margin: 0 7px;
-
-	> div {
-		position: absolute;
-		background-color: #d80;
-		height: 100%;
-	}
-}
-
-.waveformContainer {
-	position: relative;
-	width: calc(100% - 14px);
-	margin: 0 7px;
-}
-
-.waveform {
-	width: 100%;
-	max-height: 100px;
-}
-
-.waveformMarker {
-	width: 1px;
-	height: 100%;
-	background: #dd8800a0;
-	position: absolute;
-	top: 0;
-}
-
-.notesToEditor {
-	border: 1px solid #666;
-	background-color: #125;
-	padding: 2px;
-	white-space: pre;
-}
-
-.categoryNotes {
-	border: 1px solid #666;
-	background-color: #500;
-	padding: 2px;
-	white-space: pre;
-}
-
-.chaptersEnabledSelection {
-	margin-top: 5px;
-}
-
-.transitionLabel {
-	margin-right: 5px;
-}
-
-.rangeRegion {
-	margin-top: 4px;
-}
-
-.timeRangeEntryRow {
-	display: flex;
-	gap: 4px;
-	align-items: center;
-	margin: 4px 0;
-}
-
-.timeRangeSelect {
-	flex-basis: 100px;
-
-	input {
-		width: calc(100% - 8px); // Subtract the browser-added padding and border
-	}
-}
-
-.timeRangeIcon {
-	flex-basis: 16px;
-
-	img {
-		width: 16px;
-	}
-}
