+++ thrimbletrimmer/src/editor/NotesToEditor.module.scss
@@ -1,0 +1,6 @@
+.notesToEditor {
+	border: 1px solid #666;
+	background-color: #125;
+	padding: 2px;
+	white-space: pre;
+}
