+++ thrimbletrimmer/src/editor/Waveform.module.scss
@@ -1,0 +1,18 @@
+.waveformContainer {
+	position: relative;
+	width: calc(100% - 14px);
+	margin: 0 7px;
+}
+
+.waveform {
+	width: 100%;
+	max-height: 100px;
+}
+
+.waveformMarker {
+	width: 1px;
+	height: 100%;
+	background: #dd8800a0;
+	position: absolute;
+	top: 0;
+}
